Module turbopack_ecmascript::analyzer

Modules

builtin

graph

imports

linker

top_level_await

well_known

Structs

ConstantNumber

DefineableNameIter

ModuleValue

RequireContextOptions

RequireContextValue

RequireContextValueVc

Vc for RequireContextValue

Enums

BinaryOperator

ConstantString

ConstantValue

JsValue

TODO: Use Arc There are 4 kinds of values: Leaves, Nested, Operations, and Placeholders (see [JsValueMetaKind] for details). Values are processed in two phases:

LogicalOperator

ObjectPart

PositiveBinaryOperator

WellKnownFunctionKind

A list of well-known functions that have special meaning in the analysis.

WellKnownObjectKind

A list of well-known objects that have special meaning in the analysis.

Functions

parse_require_context

Parse the arguments passed to a require.context invocation, validate them and convert them to the appropriate rust values.

Type Definitions

RequireContextValueReadRef

see [turbo_tasks::ReadRef]