Enum turbopack_ecmascript::analyzer::JsValue

pub enum JsValue {

    Constant(ConstantValue),
    Url(Url),
    WellKnownObject(WellKnownObjectKind),
    WellKnownFunction(WellKnownFunctionKind),
    Unknown(Option<Arc<JsValue>>, &'static str),
    Array {
        total_nodes: usize,
        items: Vec<JsValue>,
        mutable: bool,
    },
    Object {
        total_nodes: usize,
        parts: Vec<ObjectPart>,
        mutable: bool,
    },
    Alternatives(usize, Vec<JsValue>),
    Function(usize, u32, Box<JsValue>),
    Concat(usize, Vec<JsValue>),
    Add(usize, Vec<JsValue>),
    Not(usize, Box<JsValue>),
    Logical(usize, LogicalOperator, Vec<JsValue>),
    Binary(usize, Box<JsValue>, BinaryOperator, Box<JsValue>),
    Call(usize, Box<JsValue>, Vec<JsValue>),
    MemberCall(usize, Box<JsValue>, Box<JsValue>, Vec<JsValue>),
    Member(usize, Box<JsValue>, Box<JsValue>),
    Variable(Id),
    Argument(u32, usize),
    FreeVar(JsWord),
    Module(ModuleValue),

}

TODO: Use Arc There are 4 kinds of values: Leaves, Nested, Operations, and Placeholders (see [JsValueMetaKind] for details). Values are processed in two phases:

Analyze phase: We convert AST into JsValues. We don’t have contextual information so we need to insert placeholders to represent that.

Link phase: We try to reduce a value to a constant value. The link phase has 5 substeps that are executed on each node in the graph depth-first. When a value is modified, we need to visit the new children again.

Replace variables with their values. This replaces JsValue::Variable. No variable should be remaining after that.

Replace placeholders with contextual information. This usually replaces JsValue::FreeVar and JsValue::Module. Some JsValue::Call on well- known functions might also be replaced. No free vars or modules should be remaining after that.

Replace operations on well-known objects and functions. This handles JsValue::Call and JsValue::Member on well-known objects and functions.

Replace all built-in functions with their values when they are compile-time constant.

For optimization, any nested operations are replaced with JsValue::Unknown. So only one layer of operation remains. Any remaining operation or placeholder can be treated as unknown.

Variants

Constant(ConstantValue)

A constant primitive value.

Url(Url)

An constant URL object.

WellKnownObject(WellKnownObjectKind)

Some kind of well-known object (must not be an array, otherwise Array.concat needs to be changed)

WellKnownFunction(WellKnownFunctionKind)

Some kind of well-known function

Unknown(Option<Arc<JsValue>>, &'static str)

Not-analyzable value. Might contain the original value for additional info. Has a reason string for explanation.

Array

Fields

total_nodes: usize

items: Vec<JsValue>

mutable: bool

An array of nested values

Object

Fields

total_nodes: usize

parts: Vec<ObjectPart>

mutable: bool

An object of nested values

Alternatives(usize, Vec<JsValue>)

A list of alternative values

Function(usize, u32, Box<JsValue>)

A function reference. The return value might contain JsValue::Argument placeholders that need to be replaced when calling this function. (total_node_count, func_ident, return_value)

Concat(usize, Vec<JsValue>)

A string concatenation of values. foo.${unknownVar}.js => ‘foo’ + Unknown + ‘.js’

Add(usize, Vec<JsValue>)

An addition of values. This can be converted to JsValue::Concat if the type of the variable is string.

Not(usize, Box<JsValue>)

Logical negation !expr

Logical(usize, LogicalOperator, Vec<JsValue>)

Logical operator chain e. g. expr && expr

Binary(usize, Box<JsValue>, BinaryOperator, Box<JsValue>)

Binary expression e. g. expr == expr

Call(usize, Box<JsValue>, Vec<JsValue>)

A function call without a this context. (total_node_count, callee, args)

MemberCall(usize, Box<JsValue>, Box<JsValue>, Vec<JsValue>)

A function call with a this context. (total_node_count, obj, prop, args)

Member(usize, Box<JsValue>, Box<JsValue>)

A member access obj[prop] (total_node_count, obj, prop)

Variable(Id)

A reference to a variable.

Argument(u32, usize)

A reference to an function argument. (func_ident, arg_index)

FreeVar(JsWord)

A reference to a free variable.

Module(ModuleValue)

This is a reference to a imported module.

Implementations

impl JsValue

pub fn alternatives(list: Vec<JsValue>) -> Self

pub fn concat(list: Vec<JsValue>) -> Self

pub fn add(list: Vec<JsValue>) -> Self

pub fn logical_and(list: Vec<JsValue>) -> Self

pub fn logical_or(list: Vec<JsValue>) -> Self

pub fn nullish_coalescing(list: Vec<JsValue>) -> Self

pub fn equal(a: JsValue, b: JsValue) -> Self

pub fn not_equal(a: JsValue, b: JsValue) -> Self

pub fn strict_equal(a: JsValue, b: JsValue) -> Self

pub fn strict_not_equal(a: JsValue, b: JsValue) -> Self

pub fn logical_not(inner: Box<JsValue>) -> Self

pub fn array(items: Vec<JsValue>) -> Self

pub fn frozen_array(items: Vec<JsValue>) -> Self

pub fn function(func_ident: u32, return_value: Box<JsValue>) -> Self

pub fn object(list: Vec<ObjectPart>) -> Self

pub fn frozen_object(list: Vec<ObjectPart>) -> Self

pub fn call(f: Box<JsValue>, args: Vec<JsValue>) -> Self

pub fn member_call(o: Box<JsValue>, p: Box<JsValue>, args: Vec<JsValue>) -> Self

pub fn member(o: Box<JsValue>, p: Box<JsValue>) -> Self

impl JsValue

pub fn total_nodes(&self) -> usize

pub fn debug_assert_total_nodes_up_to_date(&mut self)

pub fn ensure_node_limit(&mut self, limit: usize)

impl JsValue

pub fn explain_args( args: &[JsValue], depth: usize, unknown_depth: usize ) -> (String, String)

pub fn explain(&self, depth: usize, unknown_depth: usize) -> (String, String)

impl JsValue

pub fn make_unknown(&mut self, reason: &'static str)

Convert the value into unknown with a specific reason.

pub fn make_unknown_without_content(&mut self, reason: &'static str)

Convert the value into unknown with a specific reason, but don’t retain the original value.

pub fn make_nested_operations_unknown(&mut self) -> bool

Make all nested operations unknown when the value is an operation.

pub fn add_unknown_mutations(&mut self)

impl JsValue

pub fn get_defineable_name_len(&self) -> Option<usize>

pub fn iter_defineable_name_rev(&self) -> DefineableNameIter<'_>

impl JsValue

pub fn as_str(&self) -> Option<&str>

Returns the constant string if the value represents a constant string.

pub fn is_truthy(&self) -> Option<bool>

Checks if the value is truthy. Returns None if we don’t know. Returns Some if we know if or if not the value is truthy.

pub fn is_falsy(&self) -> Option<bool>

Checks if the value is falsy. Returns None if we don’t know. Returns Some if we know if or if not the value is falsy.

pub fn is_nullish(&self) -> Option<bool>

Checks if the value is nullish (null or undefined). Returns None if we don’t know. Returns Some if we know if or if not the value is nullish.

pub fn is_not_nullish(&self) -> Option<bool>

Checks if we know that the value is not nullish. Returns None if we don’t know. Returns Some if we know if or if not the value is not nullish.

pub fn is_empty_string(&self) -> Option<bool>

Checks if we know that the value is an empty string. Returns None if we don’t know. Returns Some if we know if or if not the value is an empty string.

pub fn is_unknown(&self) -> bool

Returns true, if the value is unknown and storing it as condition doesn’t make sense. This is for optimization purposes.

pub fn is_string(&self) -> Option<bool>

Checks if we know that the value is a string. Returns None if we don’t know. Returns Some if we know if or if not the value is a string.

pub fn starts_with(&self, str: &str) -> Option<bool>

Checks if we know that the value starts with a given string. Returns None if we don’t know. Returns Some if we know if or if not the value starts with the given string.

pub fn ends_with(&self, str: &str) -> Option<bool>

Checks if we know that the value ends with a given string. Returns None if we don’t know. Returns Some if we know if or if not the value ends with the given string.

impl JsValue

pub async fn visit_async_until_settled<'a, F, R, E>( self, visitor: &mut F ) -> Result<(Self, bool), E>where R: 'a + Future<Output = Result<(Self, bool), E>> + Send, F: 'a + FnMut(JsValue) -> R + Send,

Visit the node and all its children with a function in a loop until the visitor returns false for the node and all children

pub async fn visit_each_children_async_until_settled<'a, F, R, E>( self, visitor: &mut F ) -> Result<Self, E>where R: 'a + Future<Output = Result<(Self, bool), E>> + Send, F: 'a + FnMut(JsValue) -> R + Send,

Visit all children of the node with an async function in a loop until the visitor returns false

pub async fn visit_async<'a, F, R, E>( self, visitor: &mut F ) -> Result<(Self, bool), E>where R: 'a + Future<Output = Result<(Self, bool), E>>, F: 'a + FnMut(JsValue) -> R,

Visit the node and all its children with an async function.

pub async fn visit_each_children_async<'a, F, R, E>( self, visitor: &mut F ) -> Result<(Self, bool), E>where R: 'a + Future<Output = Result<(Self, bool), E>>, F: 'a + FnMut(JsValue) -> R,

Visit all children of the node with an async function.

pub async fn for_each_children_async<'a, F, R, E>( self, visitor: &mut F ) -> Result<(Self, bool), E>where R: 'a + Future<Output = Result<(Self, bool), E>>, F: 'a + FnMut(JsValue) -> R,

Call an async function for each child of the node.

pub fn visit_mut_until_settled( &mut self, visitor: &mut impl FnMut(&mut JsValue) -> bool )

Visit the node and all its children with a function in a loop until the visitor returns false

pub fn visit_mut( &mut self, visitor: &mut impl FnMut(&mut JsValue) -> bool ) -> bool

Visit the node and all its children with a function.

pub fn visit_mut_conditional( &mut self, condition: impl Fn(&JsValue) -> bool, visitor: &mut impl FnMut(&mut JsValue) -> bool ) -> bool

Visit all children of the node with a function. Only visits nodes where the condition is true.

pub fn for_each_children_mut( &mut self, visitor: &mut impl FnMut(&mut JsValue) -> bool ) -> bool

Calls a function for each child of the node. Allows mutating the node. Updates the total nodes count after mutation.

pub fn for_each_early_children_mut( &mut self, visitor: &mut impl FnMut(&mut JsValue) -> bool ) -> bool

Calls a function for only early children. Allows mutating the node. Updates the total nodes count after mutation.

pub fn for_each_late_children_mut( &mut self, visitor: &mut impl FnMut(&mut JsValue) -> bool ) -> bool

Calls a function for only late children. Allows mutating the node. Updates the total nodes count after mutation.

pub fn visit(&self, visitor: &mut impl FnMut(&JsValue))

Visit the node and all its children with a function.

pub fn for_each_children(&self, visitor: &mut impl FnMut(&JsValue))

Calls a function for all children of the node.

impl JsValue

pub fn normalize_shallow(&mut self)

Normalizes only the current node. Nested alternatives, concatenations, or operations are collapsed.

pub fn normalize(&mut self)

Normalizes the current node and all nested nodes.