Trait turbopack_ecmascript::chunk::EcmascriptChunkRuntime

pub trait EcmascriptChunkRuntime: ValueToStringwhere
    TaskInput: for<'a> From<&'a Self>,{
    // Required methods
    fn decorate_asset_ident(
        &self,
        chunk: EcmascriptChunkVc,
        ident: AssetIdentVc
    ) -> AssetIdentVc;
    fn with_chunk_group(&self, chunk_group: ChunkGroupVc) -> Self;
    fn references(&self, chunk: EcmascriptChunkVc) -> AssetReferencesVc;
    fn content(
        &self,
        chunk: EcmascriptChunkVc
    ) -> EcmascriptChunkRuntimeContentVc;
    fn merge(
        &self,
        runtimes: Vec<EcmascriptChunkRuntimeVc>
    ) -> EcmascriptChunkRuntimeVc;
}

The runtime for an EcmaScript chunk.

Required Methods

fn decorate_asset_ident( &self, chunk: EcmascriptChunkVc, ident: AssetIdentVc ) -> AssetIdentVc

Decorates the asset identifiers of the chunk to make it unique for this runtime.

fn with_chunk_group(&self, chunk_group: ChunkGroupVc) -> Self

Sets a custom chunk group for this runtime. This is used in the optimizer.

fn references(&self, chunk: EcmascriptChunkVc) -> AssetReferencesVc

Returns references from this runtime.

fn content(&self, chunk: EcmascriptChunkVc) -> EcmascriptChunkRuntimeContentVc

Returns the content of the given chunk adapted to this runtime.

fn merge( &self, runtimes: Vec<EcmascriptChunkRuntimeVc> ) -> EcmascriptChunkRuntimeVc

Merges this runtime with the given runtimes.