Module turbopack_ecmascript::analyzer

Modules

builtin

graph

imports

linker

well_known

Structs

ConstantNumber

DefineableNameIter

ModuleValue

Enums

BinaryOperator

ConstantString

ConstantValue

JsValue

TODO: Use Arc There are 4 kinds of values: Leaves, Nested, Operations, and Placeholders (see [JsValueMetaKind] for details). Values are processed in two phases:

LogicalOperator

ObjectPart

PositiveBinaryOperator

WellKnownFunctionKind

A list of well-known functions that have special meaning in the analysis.

WellKnownObjectKind

A list of well-known objects that have special meaning in the analysis.